refactor: add common exceptions and base_scraper
