+++ parsers/features/exceptions.py
@@ -1,0 +1,2 @@
+class ScrapingError(Exception):
+    pass
