+++ parsers/features/schedule/scrapers.py
@@ -1,54 +1,9 @@
-from abc import ABCMeta, abstractmethod
-from functools import wraps
-from typing import TypeVar
-
 from bs4 import BeautifulSoup
 from dtos import GroupIdDTO
 from endpoints import get_search_groups_url
+from features.base_scraper import BaseScraper
 from furl import furl
 from infra.client import Client, configs
-
-T = TypeVar("T")
-
-
-class ScrapingError(Exception):
-    pass
-
-
-class CallWrapABCMeta(ABCMeta):
-    def __new__(mcls, name, bases, namespace, **kwargs) -> type:
-        cls = super().__new__(mcls, name, bases, namespace, **kwargs)
-        mcls._maybe_wrap_call(cls)
-        return cls
-
-    @staticmethod
-    def _maybe_wrap_call(cls: type) -> None:
-        call_func = cls.__dict__.get("__call__", None)
-        if call_func is None:
-            return
-        if getattr(call_func, "__isabstractmethod__", False):
-            return
-        if getattr(call_func, "__wrapped_by_meta__", False):
-            return
-
-        @wraps(call_func)
-        def wrapper(*args, **kwargs):
-            try:
-                return call_func(*args, **kwargs)
-            except Exception as err:
-                raise ScrapingError(str(err)) from err
-
-        wrapper.__wrapped_by_meta__ = True
-        cls.__call__ = wrapper
-
-
-class BaseScraper[T](metaclass=CallWrapABCMeta):
-    def __init__(self, client: Client) -> None:
-        self._client = client
-
-    @abstractmethod
-    def __call__(self, *args, **kwargs) -> T:
-        raise NotImplementedError
 
 
 class GroupIdScraper(BaseScraper[GroupIdDTO]):
@@ -73,7 +28,7 @@ if __name__ == "__main__":
     )
     with Client(configs1) as main_client:
         scraper = GroupIdScraper(main_client)
-        for _i in range(3):
-            group_id_dto = scraper("513090240003")
+        for _i in range(30):
+            group_id_dto = scraper("5130902/40003")
             print(group_id_dto)
             print()
